Update README.md
I 'have started a curation website for Software Testing space. I wasn't sure as to which category to put this in. Hence, have added to the Others section. If this is not suitable section then please let me know where to add it and I'll make necessary changes and submit it again. :)